+++ vantage6-server/vantage6/server/model/run.py
@@ -1,7 +1,7 @@
 import datetime
 import logging
 
-from sqlalchemy import Column, Text, DateTime, Integer, ForeignKey
+from sqlalchemy import Column, Text, DateTime, Integer, ForeignKey, Boolean
 from sqlalchemy.orm import relationship
 from sqlalchemy.orm.exc import NoResultFound, MultipleResultsFound
 
@@ -49,6 +49,9 @@ class Run(Base):
         Organization that executed the task
     ports : list[:class:`.~vantage6.server.model.algorithm_port.AlgorithmPort`]
         List of ports that are part of this result
+    blob_storage_used : bool
+        Whether blob storage is used for the input and result data
+        Defaults to False
     """
 
     # fields
@@ -62,6 +65,9 @@ class Run(Base):
     status = Column(Text)
     log = Column(Text)
     cleanup_at = Column(DateTime, nullable=True)
+    blob_storage_used = Column(
+        "blob_storage_used", Boolean, default=False, nullable=True
+    )
 
     # relationships
     task = relationship("Task", back_populates="runs")