+++ vantage6-server/vantage6/server/resource/node.py
@@ -134,8 +134,8 @@ node_input_schema = NodeInputSchema()
 
 
 class NodeBase(ServicesResources):
-    def __init__(self, socketio, mail, api, permissions, config):
-        super().__init__(socketio, mail, api, permissions, config)
+    def __init__(self, socketio, storage_adapter, mail, api, permissions, config):
+        super().__init__(socketio, storage_adapter, mail, api, permissions, config)
         self.r: RuleCollection = getattr(self.permissions, module_name)