+++ vantage6-server/vantage6/server/resource/organization.py
@@ -107,8 +107,8 @@ org_input_schema = OrganizationInputSchema()
 
 
 class OrganizationBase(ServicesResources):
-    def __init__(self, socketio, mail, api, permissions, config):
-        super().__init__(socketio, mail, api, permissions, config)
+    def __init__(self, socketio, storage_adapter, mail, api, permissions, config):
+        super().__init__(socketio, storage_adapter, mail, api, permissions, config)
         self.r: RuleCollection = getattr(self.permissions, module_name)