+++ vantage6-server/vantage6/server/resource/run.py
@@ -3,6 +3,7 @@ from typing import Union
 import sqlalchemy as sa
 
 from flask import g, request
+
 from flask_restful import Api
 from http import HTTPStatus
 from sqlalchemy import desc
@@ -130,8 +131,8 @@ def permissions(permissions: PermissionManager):
 class RunBase(ServicesResources):
     """Base class for run resources"""
 
-    def __init__(self, socketio, mail, api, permissions, config):
-        super().__init__(socketio, mail, api, permissions, config)
+    def __init__(self, socketio, storage_adapter, mail, api, permissions, config):
+        super().__init__(socketio, storage_adapter, mail, api, permissions, config)
         self.r: RuleCollection = getattr(self.permissions, module_name)