+++ vantage6-server/vantage6/server/resource/user.py
@@ -105,8 +105,8 @@ user_schema_with_permissions = UserWithPermissionDetailsSchema()
 
 
 class UserBase(ServicesResources):
-    def __init__(self, socketio, mail, api, permissions, config):
-        super().__init__(socketio, mail, api, permissions, config)
+    def __init__(self, socketio, storage_adapter, mail, api, permissions, config):
+        super().__init__(socketio, storage_adapter, mail, api, permissions, config)
         self.r: RuleCollection = getattr(self.permissions, module_name)