+++ vantage6/vantage6/cli/node/clean.py
@@ -43,4 +43,6 @@ def cli_node_clean() -> None:
                 )
                 debug(e)
                 exit(1)
-    info("Done!")
+        info("Temporary Docker volumes removed successfully.")
+    else:
+        info("Cancelled, no volumes were removed.")